packages/casbin-fastapi-decorator-db/pyproject.toml
[project]
name = "casbin-fastapi-decorator-db"
version = "0.1.0"
description = "Database enforcer provider for casbin-fastapi-decorator"
requires-python = ">=3.10"
dependencies = [
    "casbin-fastapi-decorator",
    "sqlalchemy>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
casbin-fastapi-decorator = { workspace = true }
